Extract utils functions from confirmation-popup and search-query
