+++ integreat_cms/static/src/js/media-management/library.tsx
@@ -15,7 +15,7 @@ import EditSidebar from "./component/edit-sidebar";
 import { Message } from "./component/message";
 import CreateDirectory from "./component/create-directory";
 import UploadFile from "./component/upload-file";
-import { setSearchQueryEventListeners } from "../feature/search-query";
+import { setSearchQueryEventListeners } from "../utils/search-query-listeners";
 import { getCsrfToken } from "../utils/csrf-token";
 
 export type LibraryProps = {
