+++ integreat_cms/static/src/js/pages/fetch-subpages.ts
@@ -7,7 +7,7 @@ import { getCsrfToken } from "../utils/csrf-token";
 import { setBulkActionEventListeners } from "../bulk-actions";
 import { setToggleSubpagesEventListeners } from "./toggle-subpages";
 import { addDragAndDropListeners } from "../tree-drag-and-drop";
-import { addConfirmationDialogListeners } from "../confirmation-popups";
+import { addConfirmationDialogListeners } from "../utils/confirmation-popup";
 import { addPreviewWindowListeners, openPreviewWindowInPageTree } from "./page-preview";
 import { restorePageTreeLayout } from "./persistent_page_tree";
 import { setPageAccessesEventListeners } from "../analytics/statistics-page-accesses";
